+++ src/main/java/com/heyyoung/solsol/domain/menu/dto/GetMenuResponse.java
@@ -4,7 +4,7 @@ import com.heyyoung.solsol.domain.menu.MenuEntity;
 
 import java.math.BigDecimal;
 
-public record GetMenuResponse(int menuId, String name, BigDecimal price, String image) {
+public record GetMenuResponse(long menuId, String name, BigDecimal price, String image) {
     public static GetMenuResponse from(MenuEntity menu) {
         return new GetMenuResponse(menu.getMenuId(), menu.getName(), menu.getPrice(), menu.getImage());
     }
